+++ src/plume/PI_SourceGeometry_Cube.hpp
@@ -46,12 +46,12 @@ private:
   // note that this also inherits public data members ReleaseType* m_rType and SourceShape m_sShape.
   // guidelines for how to set these variables within an inherited source are given in SourceType.
 
-  double m_minX = -1.0;
-  double m_minY = -1.0;
-  double m_minZ = -1.0;
-  double m_maxX = -1.0;
-  double m_maxY = -1.0;
-  double m_maxZ = -1.0;
+  float m_minX = -1.0;
+  float m_minY = -1.0;
+  float m_minZ = -1.0;
+  float m_maxX = -1.0;
+  float m_maxY = -1.0;
+  float m_maxZ = -1.0;
 
   std::random_device rd;// Will be used to obtain a seed for the random number engine
   std::mt19937 prng;// Standard mersenne_twister_engine seeded with rd()
@@ -71,21 +71,21 @@ public:
 
   void parseValues() override
   {
-    parsePrimitive<double>(true, m_minX, "minX");
-    parsePrimitive<double>(true, m_minY, "minY");
-    parsePrimitive<double>(true, m_minZ, "minZ");
-    parsePrimitive<double>(true, m_maxX, "maxX");
-    parsePrimitive<double>(true, m_maxY, "maxY");
-    parsePrimitive<double>(true, m_maxZ, "maxZ");
+    parsePrimitive<float>(true, m_minX, "minX");
+    parsePrimitive<float>(true, m_minY, "minY");
+    parsePrimitive<float>(true, m_minZ, "minZ");
+    parsePrimitive<float>(true, m_maxX, "maxX");
+    parsePrimitive<float>(true, m_maxY, "maxY");
+    parsePrimitive<float>(true, m_maxZ, "maxZ");
   }
 
 
-  void checkPosInfo(const double &domainXstart,
-                    const double &domainXend,
-                    const double &domainYstart,
-                    const double &domainYend,
-                    const double &domainZstart,
-                    const double &domainZend) override;
+  void checkPosInfo(const float &domainXstart,
+                    const float &domainXend,
+                    const float &domainYstart,
+                    const float &domainYend,
+                    const float &domainZstart,
+                    const float &domainZend) override;
 
   void setInitialPosition(vec3 &) override;
 };
