+++ src/plume/PI_SourceGeometry_Point.cpp
@@ -39,7 +39,12 @@
 #define _USE_MATH_DEFINES
 #include <cmath>
 
-void PI_SourceGeometry_Point::checkPosInfo(const double &domainXstart, const double &domainXend, const double &domainYstart, const double &domainYend, const double &domainZstart, const double &domainZend)
+void PI_SourceGeometry_Point::checkPosInfo(const float &domainXstart,
+                                           const float &domainXend,
+                                           const float &domainYstart,
+                                           const float &domainYend,
+                                           const float &domainZstart,
+                                           const float &domainZend)
 {
   if (posX < domainXstart || posX > domainXend) {
     std::cerr << "[ERROR] \t PI_SourceGeometry_Point::checkPosInfo: \n\t\t input posX is outside of domain! posX = \"" << posX
